requirements.txt
asgiref==3.3.4
backports.zoneinfo==0.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-cors-headers==3.10.1
django-restframework==0.0.1
djangorestframework==3.12.4
filter==0.0.0.20200724
idna==2.10
importlib-metadata==4.8.3
Markdown==3.3.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
tzdata==2021.5
urllib3==1.26.4
zipp==3.6.0
